A vector built from roots is one source, not three copies
`map_entry_as_vec` and `vec_from_roots` were the last pair in `vector.rs`
that every runtime wrote out by hand -- six lines each, three times, and
nothing compared them. `kin/vecroots.kin` generates both. That is the fourth
and last of the `(base, n)` constructors, after `list-from-roots`,
`b-from-roots` and `rope-from-roots`.

`map-entry-as-vec` came OFF the vocabulary in the same change. It was declared
in `flint.impl.rt` as a call into three hand-written bodies; a source that
wants it now requires it like any other generated function, and the warning
that it is NOT `map-entry-vector` moved to where a reader of the definition
will see it. The shim on `Vec` had to go with it on both ports: a generated
module star-imports both `Vec` and `Vecroots`, and two statics of one name make
every call AMBIGUOUS -- javac refuses `Collwrite.java` outright. `Builtins`
says the generated name now.

`n` is `u32` rather than `usize`, which is what the tag `I32` means and what
`list_from_roots` beside it already took. Four call sites gained `as u32`,
including the one in `codec.rs` that was already writing `n as u32` on the
very next line for the list.

WHAT THIS DID NOT PORT, and why, because the count is misleading. The four
files hold 55 functions; `bytes.rs` is 29 of them, which is where "about
twenty-nine left" comes from, and every one is a vocabulary primitive holding
a host `Vec<u8>`, a host `&str`, a borrowed slice or the allocator. Those are
named three ways each in `flint.impl.rt`, so they are the opposite of drift.
`strs.rs` is the same wall. `is_vector` is `is_heap_ty(v, TY_VEC)` spelled
out and both ports already say the latter, so there is no second
implementation to converge -- generating it would add a `pub` function to
three modules to give one runtime's alias a home.

`pike.rs` IS THE ONE THAT SHOULD BE GENERATED AND IS NOT, and
`DECISIONS.md#the-pike-vm-is-the-last-triplicate` records why: kin renders
`^:mut` on a parameter as a by-value rebinding in Rust and a shared reference
on both ports, so no tag spells a mutable host array in all three. Reading the
three copies found five divergences no gate compares. One is fixed here: both
ports narrowed `re-run`'s `from` to `int` BEFORE clamping it, so 2^31 became
-2147483648, passed the length guard and indexed the code points negatively --
guest code throwing an out-of-bounds out of the host. Native answers nil. The
other four are recorded, including a hang reachable from a crafted program,
which is left open because the fix is a decision about `re-compile`'s contract
rather than a bug fix.

Ran: `kin/scripts/verify` on vecroots and collwrite; `bin/check-kin` (90
sources, all agree, every module reached); `bin/check-decisions`;
`bin/check-builtins`; `bin/check-flattens`; `cargo test -p flint-rt` (137
tests, 0 failures); `bb test/{types,bytes,ropes,regex_pike,threads,twobuilds}.clj`;
javac over the whole JVM runtime and `dotnet build` over the CLR one. The
consolidated `bin/test`/`bin/conform-hosts` gate was NOT run here.

+++ runtime/src/codec.rs
@@ -500,7 +500,7 @@ impl Rt {
                     self.push(v);
                 }
                 let out = match tag {
-                    K_VECTOR => self.vec_from_roots(base, n),
+                    K_VECTOR => self.vec_from_roots(base, n as u32),
                     K_LIST => self.list_from_roots(base, n as u32),
                     _ => {
                         // No `set_from_roots`, so it is built by conj -- and
